+++ src/pages/profile/index.tsx
@@ -5,11 +5,11 @@ import { useRouter } from "next/router";
 import { useState } from "react";
 import { useForm, SubmitHandler } from "react-hook-form";
 import { useSWRConfig } from "swr";
-import { authOptions } from "./api/auth/[...nextauth]";
-import { getProfile } from "./api/user";
-import { updateUser } from "../hooks";
-import { UserProfileForm, UserProfileValues } from "../components";
-import { User } from "../types";
+import { authOptions } from "../api/auth/[...nextauth]";
+import { getProfile } from "../api/user";
+import { updateUser } from "../../hooks";
+import { UserProfileForm, UserProfileValues } from "../../components";
+import { User } from "../../types";
 
 type Props = {
   user: User;